+++ src/main/java/com/seq/api/Transaction.java
@@ -66,6 +66,7 @@ public class Transaction {
   /**
    * Iterable interface for consuming pages of transactions from a query.
    */
+  @Deprecated
   public static class PageIterable extends BasePageIterable<Page> {
     public PageIterable(Client client, String path, Query nextQuery) {
       super(client, path, nextQuery, Page.class);
@@ -117,7 +118,9 @@ public class Transaction {
      * @param client ledger API connection object
      * @return an iterable over pages of transactions
      * @throws ChainException
+     * @deprecated use {@link #getPage} instead
      */
+    @Deprecated
     public PageIterable getPageIterable(Client client) throws ChainException {
       return new PageIterable(client, "list-transactions", this.next);
     }