+++ CHANGELOG.md
@@ -10,6 +10,11 @@
   property is `tags`.
 * The `assetAlias` and `assetId` properties on `ActionSum` have been deprecated;
   the new property is `flavorId`.
+* `BasePageIterable` and its subclasses `Account.PageIterable`,
+  `ActionSum.PageIterable`, `Feed.PageIterable`, `Flavor.PageIterable`,
+  `Key.PageIterable`, and `Transaction.PageIterable` have been deprecated;
+  seee [Pagination](https://dashboard.seq.com/docs/pagination) for new
+  interfaces.
 
 ## 1.5.1 (20180320)